# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRIS Team 4 API Documentation",
        "description": "Dokumentasi API aplikasi HRIS Untuk PBL Team 4",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/admin/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register admin baru",
                "operationId": "8300ba06ee0c588b326e6db715f234d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login sebagai admin",
                "operationId": "6fd02a4c8e5d0723bf3d1954ccedb6f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    },
                    "422": {
                        "description": "Login gagal atau data tidak valid"
                    }
                }
            }
        },
        "/admin/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil data admin yang login",
                "operationId": "765d0a6c0bb0b9a0f6757751f2c00959",
                "responses": {
                    "200": {
                        "description": "Data admin berhasil diambil"
                    },
                    "403": {
                        "description": "Bukan admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/employee/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login karyawan",
                "operationId": "78d40c59d00e7d974e1470fd7961fada",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company",
                                    "employee_id",
                                    "password"
                                ],
                                "properties": {
                                    "company": {
                                        "type": "string"
                                    },
                                    "employee_id": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    },
                    "401": {
                        "description": "Login gagal"
                    },
                    "404": {
                        "description": "Perusahaan tidak ditemukan"
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user (admin atau employee)",
                "description": "Logout untuk user yang sudah login, baik admin maupun employee, menggunakan token sanctum.",
                "operationId": "5490f6f73efe10a780296f44e62806ba",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil data user yang sedang login (admin atau employee)",
                "description": "Mengambil data user yang sedang login, baik admin maupun employee, berdasarkan token sanctum.",
                "operationId": "b0a66bffa8f791f745d551c2f01a642c",
                "responses": {
                    "200": {
                        "description": "Data user berhasil diambil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/company": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Membuat data perusahaan baru",
                "operationId": "cc66eb61ef990224e42be56da6d61e29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "company_username"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "example": "PT Maju Mundur"
                                    },
                                    "company_username": {
                                        "type": "string",
                                        "example": "majumundur123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company berhasil dibuat.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Company berhasil dibuat."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "c2c1a7d0-3b8d-4e3d-a7e2-123456789abc"
                                                },
                                                "company_name": {
                                                    "type": "string",
                                                    "example": "PT Maju Mundur"
                                                },
                                                "company_username": {
                                                    "type": "string",
                                                    "example": "majumundur123"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The company_username has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mengirimkan link reset password ke email pengguna",
                "operationId": "b938c14112bf8a2908f62fbba36b6d3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link reset password berhasil dikirim"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Gagal mengirim email reset password"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mengatur ulang password menggunakan token reset",
                "operationId": "11caca5894ad4fb1f89b0bf3a24f6d0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "reset-token-here"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecurepassword"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecurepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil direset"
                    },
                    "400": {
                        "description": "Token tidak valid atau data salah"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/reset-password/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Menampilkan token reset password",
                "operationId": "0565640e76bdaa2feba5d44d6632689d",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token reset password dari email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token berhasil ditampilkan"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Company",
            "description": "Company"
        }
    ]
}